message osd before closing

+++ acl/authentifier.hpp
@@ -231,7 +231,7 @@ public:
             return true;
         }
 
-        const long enddate = this->ini.context.end_date_cnx.get();
+        const uint32_t enddate = this->ini.context.end_date_cnx.get();
         if (enddate != 0 && (now > enddate)) {
             LOG(LOG_INFO, "Session is out of allowed timeframe : closing");
             const char * message = "Session is out of allowed timeframe";
